Changed architecture. Added a separate auth application
