+++ backend/apps/auth_app/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class AuthConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.auth_app'
+
+    def ready(self):
+        import apps.auth_app.signals
