+++ backend/apps/home/apps.py
@@ -4,6 +4,3 @@ from django.apps import AppConfig
 class HomeAppConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'apps.home'
-
-    def ready(self):
-        import apps.home.signals
