+++ backend/apps/home/utils.py
@@ -1,5 +1,7 @@
-from collections import defaultdict
 import pandas as pd
+
+from collections import defaultdict
+
 from .models import (
     Olimpiad, BaseOlimpiadLesson, AdditionalOlimpiadLesson, OlimpiadSS,
     SoftSkill, Lesson
